+++ packages/polaris-viz/src/components/SimpleNormalizedChart/Chart.tsx
@@ -3,16 +3,13 @@ import {sum} from 'd3-array';
 import {scaleLinear} from 'd3-scale';
 import {
   COLOR_VISION_SINGLE_ITEM,
+  DataSeries,
   DEFAULT_THEME_NAME,
 } from '@shopify/polaris-viz-core';
-import type {
-  DataPoint,
-  Direction,
-  LabelFormatter,
-} from '@shopify/polaris-viz-core';
+import type {Direction, LabelFormatter} from '@shopify/polaris-viz-core';
 
 import type {ComparisonMetricProps} from '../ComparisonMetric';
-import {getSeriesColors} from '../../hooks/useThemeSeriesColors';
+import {useThemeSeriesColors} from '../../hooks/useThemeSeriesColors';
 import {
   useColorVisionEvents,
   usePrefersReducedMotion,
@@ -26,7 +23,7 @@ import type {Size, LabelPosition} from './types';
 import styles from './SimpleNormalizedChart.scss';
 
 export interface ChartProps {
-  data: DataPoint[];
+  data: DataSeries[];
   comparisonMetrics?: Omit<ComparisonMetricProps, 'theme'>[];
   labelFormatter?: LabelFormatter;
   labelPosition?: LabelPosition;
@@ -44,11 +41,16 @@ export function Chart({
   size = 'small',
   theme = DEFAULT_THEME_NAME,
 }: ChartProps) {
+  const flattenedData = data.map(({data}) => data).flat();
+
   useColorVisionEvents();
 
   const selectedTheme = useTheme(theme);
-  const colors = getSeriesColors(data.length, selectedTheme);
-  const containsNegatives = data.some(({value}) => value !== null && value < 0);
+  const colors = useThemeSeriesColors(data, selectedTheme);
+
+  const containsNegatives = flattenedData.some(
+    ({value}) => value !== null && value < 0,
+  );
   const isDevelopment = process.env.NODE_ENV === 'development';
   const {prefersReducedMotion} = usePrefersReducedMotion();
 
@@ -66,13 +68,13 @@ export function Chart({
     );
   }
 
-  if (isDevelopment && data.length > 4) {
+  if (isDevelopment && flattenedData.length > 4) {
     throw new Error(
       'This component displays a max of 4 data items. Please modify your data before passing it into this component.',
     );
   }
 
-  const slicedData = data.slice(0, 4);
+  const slicedData = flattenedData.slice(0, 4);
   const totalValue = sum(slicedData, ({value}) => value);
 
   const xScale = scaleLinear().range([0, 100]).domain([0, totalValue]);
@@ -119,8 +121,8 @@ export function Chart({
             <BarLabel
               activeIndex={activeIndex}
               index={index}
-              key={`${key}-${formattedValue}`}
-              label={`${key}`}
+              key={`${key}-${formattedValue}-${index}`}
+              label={`${data[index].name}`}
               value={formattedValue}
               color={colors[index]}
               comparisonMetric={comparisonMetric}
@@ -155,7 +157,7 @@ export function Chart({
               direction={direction}
               size={size}
               scale={xScale(value)}
-              key={`${key}`}
+              key={`${key}-${index}`}
               color={colors[colorIndex]}
               roundedCorners={selectedTheme.bar.hasRoundedCorners}
             />